mbuto: Add support for additional libraries typically loaded via dlopen(3)
Example: glibc needs, for basic tasks such as domain resolution,
libnss_files and libnss_dns, which in turn links to libresolv.

Add a LIBS_DLOPEN variable to hold pairs of libraries typically
doing this and additional libraries to copy -- not a complete list,
just what's strictly needed for basic functionality.

Signed-off-by: Stefano Brivio <sbrivio@redhat.com>