mbuto: add the -d switch back, but remove it from the variable
When specify -d as part of __d_opt variable, it is passed as a single string to
modprobe, like "-d /foo/bar", and modprobe fails to figure out the directory
name correctly. With this patch, we pass it as two different strings, like "-d"
"/foo/bar", and modprobe works correctly.

Signed-off-by: Sevinj Aghayeva <sevinj.aghayeva@gmail.com>